+++ admin-tool/src/pages/LoginPage.js
@@ -6,7 +6,6 @@ import Avatar from '@material-ui/core/Avatar';
 import Button from '@material-ui/core/Button';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import TextField from '@material-ui/core/TextField';
-import Link from '@material-ui/core/Link';
 import Box from '@material-ui/core/Box';
 import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
 import Typography from '@material-ui/core/Typography';
@@ -14,19 +13,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
 import { isEmail } from '../util/Validation';
 import { useCookies } from 'react-cookie';
-
-const Copyright = () => {
-  return (
-    <Typography variant="body2" color="textSecondary" align="center">
-      {'Copyright © '}
-      <Link color="inherit" href="https://capzzang.co.kr">
-        ITTABI
-      </Link>{' '}
-      2020.08
-      {'.'}
-    </Typography>
-  );
-};
+import Copyright from '../components/Copyright';
 
 const useStyles = makeStyles((theme) => ({
   paper: {
